requirements.txt
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.95.0
requests==2.31.0
google-auth-oauthlib>=1.0.05
